Force MTU 9216 on lcpng TAP interfaces
The TAP interface does copy forward the host MTU based on the VPP
interface's L3 MTU, but it should also ensure that the VPP tap
interface has an MTU that is greater-or-equal to those.  Considering
users can set the interfaces at runtime (set interface mtu packet ...)
ensure that the tap MTU is large enough.

Ideally, a callback updates the MTU to the same value as the L3 MTU of
the VPP interface, or the L3 MTU of the host interface, if either of
those change. For now, it's a safe bet to take jumbo 9216.