Fix bug in find_outer_vlan()
I was looking at the hw interface list, which makes sense for ethernet
devices but not for other devices, notably BondEthernets.

In addition, creating two separate interfaces with the same outer
(for example Gi3/0/0 dot1ad 2345 inner-dot1q 1000 AND the same on
Gi3/0/1) would yield an erratic match and a crash.

Switch to walking the sw interface list instead, and search for
the sup_sw_if_index that has the desired outer. Result:

BondEthernet0.{1234,1235,1236,1237} can be created and are functional.